requirements.txt
alembic==1.16.5
amqp==5.3.1
async-timeout==5.0.1
bcrypt==5.0.0
billiard==4.2.2
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
WTForms==3.2.1
xgboost==3.0.5
